fix(storage/row): checked conversions on the SQLite path (#289)
Every `as` cast under src/storage/row reviewed (48 sites):

- local_payload_size returns usize (its result is bounded by the u32
  usable size), removing eleven `as usize` at its call sites.
- Rowids and interior keys are stored as unsigned varints of their
  two's-complement bits; rowid_from_varint/rowid_to_varint are now the
  one documented place that reinterpretation happens (five sites).
- Untrusted sizes and offsets (record header lengths, journal/WAL file
  sizes, sqlite_master root pages, VFS offsets) go through try_from with
  a non-panicking fallback that fails the existing length/range checks
  as a typed error instead of wrapping.
- Serial-type integer bodies are written from to_be_bytes() slices like
  their 3/6-byte neighbours; the 1-byte decode uses i8::from_be_bytes.
- Counters bounded by u32 page numbers saturate explicitly.

1700 tests pass. The crate-wide storage allow comes off in a later commit
once the Parquet side is done.

+++ src/storage/row/btree/index/index_insert.rs
@@ -86,8 +86,7 @@ fn encode_index_cell(
     payload: &[u8],
 ) -> Result<Vec<u8>, BtreeError> {
     let payload_len = payload.len() as u64;
-    let local_size =
-        (local_payload_size(usable_size, payload_len, true) as usize).min(payload.len());
+    let local_size = (local_payload_size(usable_size, payload_len, true)).min(payload.len());
     let (local_bytes, overflow_bytes) = payload.split_at(local_size);
     let mut cell = encode_varint(payload_len);
     cell.extend_from_slice(local_bytes);
